[@chicaothiago/@viniciusgcardozo] Issue 60 WIP
