+++ src/pages/Items/IndexItems/IndexItems.test.js
@@ -1,20 +1,17 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import LoadingBar from '../../_common/components/LoadingBar';
-import IndexItems from './index';
+import IndexItems from '../index';
+
 
 describe('<IndexItems/>', () => {
   describe('<LoadingBar />', () => {
-    const setProgress = jest.fn();
-    const useStateSpy = jest.spyOn(React, 'useState');
-
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('should render LoadingBar with loading true', () => {
-      useStateSpy.mockImplementation(() => [true, setProgress]);
+    it('should render LoadingBar with no props be sent', () => {
       const component = render(<IndexItems />);
+      expect(LoadingBar).toBeInTheDocument();
+    });
+    it('should render LoadingBar with loading true', () => {
+      const component = render(<IndexItems loading={true} />);
       expect(LoadingBar).toBeDefined();
     });
   });
