+++ src/pages/Items/index.jsx
@@ -1,21 +1,24 @@
-import React from 'react';
-import { Container } from '@material-ui/core';
-import Layout from '../_layout/Layout';
+import React, {
+  useEffect,
+  useState,
+} from 'react';
+import IndexItems from './IndexItems/IndexItems';
 import { listaAcessorios } from '../../models/acessorio';
 import { Item } from '../../models/item';
-import ItemsPage from './ItemsPage';
 
-const IndexItems = () => {
-  const items = listaAcessorios.map((item) => Item({ nome: item }));
-  return (
-    <>
-      <Layout>
-        <Container>
-          <ItemsPage items={items} />
-        </Container>
-      </Layout>
-    </>
-  );
+
+export default function () {
+  const [items, setItems] = useState([]);
+  const [progress, setProgress] = useState(0);
+  const [loading, setLoading] = useState(true);
+
+  const handleEffect = () => {
+    setItems(listaAcessorios.map((item) => Item({ nome: item })));
+    setProgress(100);
+    setLoading(false);
+  };
+
+  useEffect(handleEffect, []);
+
+  return <IndexItems loading={loading} progress={progress} items={items} />;
 };
-
-export default IndexItems;
